refact: Moving filters to the filters module
