+++ py_search_engine/analysis/analyzers/__init__.py
@@ -1,0 +1,1 @@
+from .simple_analyzer import SimpleAnalyzer
