+++ py_search_engine/analysis/analyzers/simple_analyzer.py
@@ -1,36 +1,7 @@
-from py_search_engine.analysis.analyzers.base_analyzer import BaseAnalyzer
+from py_search_engine.analysis.filters import (SimpleNormalizeFilter,
+                                               SimpleTokenizer)
 
-
-class BaseFilter:
-    def __init__(self, input_list):
-        self.input_list = input_list
-
-    def process(self):
-        # TODO: optimize
-        processed_terms = []
-
-        for term in self.input_list:
-            processed_terms += self._process_term(term)
-
-        return processed_terms
-
-    def _process_term(self, item):
-        raise NotImplementedError
-
-    def _flatten(self, token_list):
-        return [token for token_sublist in token_list for token in token_sublist]
-
-
-class SimpleNormalizeFilter(BaseFilter):
-    def _process_term(self, term):
-        # TODO: replaces special chars...
-        # TODO: lowercases...
-        return [term]
-
-
-class SimpleTokenizer(BaseFilter):
-    def _process_term(self, term):
-        return term.split(' ')
+from .base_analyzer import BaseAnalyzer
 
 
 class SimpleAnalyzer(BaseAnalyzer):
