+++ py_search_engine/analysis/filters/__init__.py
@@ -1,0 +1,2 @@
+from .simple_normalize_filter import SimpleNormalizeFilter
+from .simple_tokenizer import SimpleTokenizer
