+++ py_search_engine/analysis/filters/simple_normalize_filter.py
@@ -1,0 +1,8 @@
+from .base_filter import BaseFilter
+
+
+class SimpleNormalizeFilter(BaseFilter):
+    def _process_term(self, term):
+        # TODO: replaces special chars...
+        # TODO: lowercases...
+        return [term]
