+++ py_search_engine/analysis/filters/simple_tokenizer.py
@@ -1,0 +1,6 @@
+from .base_filter import BaseFilter
+
+
+class SimpleTokenizer(BaseFilter):
+    def _process_term(self, term):
+        return term.split(' ')
